Detach RecentlyLikedChatHistoryList.js.js from DashboardPage.js
